// .tmp/zip_snapshot/tote-bag-practicantes/apps/api/src/generated/client/enums.ts
export const PriceRuleScope = {
  B2C: 'B2C',
  B2B: 'B2B'
} as const

export type PriceRuleScope = (typeof PriceRuleScope)[keyof typeof PriceRuleScope]